+++ blenderproc/__init__.py
@@ -1,5 +1,10 @@
 import os
 import sys
+from .version import __version__
+
+# check the python version, only python 3.X is allowed:
+if sys.version_info.major < 3:
+    raise Exception("BlenderProc requires at least python 3.X to run.")
 
 # Only import if we are in the blender environment, this environment variable is set by the run.py script
 if "INSIDE_OF_THE_INTERNAL_BLENDER_PYTHON_ENVIRONMENT" in os.environ: