+++ blenderproc/python/types/MeshObjectUtility.py
@@ -5,8 +5,12 @@ import numpy as np
 import bmesh
 import mathutils
 from mathutils import Vector, Matrix
-from blenderproc.external.vhacd.decompose import convex_decomposition
+from sys import platform
 
+
+if platform != "win32":
+    # this is only supported under linux and mac os, the import itself already doesn't work under windows
+    from blenderproc.external.vhacd.decompose import convex_decomposition
 from blenderproc.python.types.EntityUtility import Entity
 from blenderproc.python.utility.Utility import Utility, resolve_path
 from blenderproc.python.utility.BlenderUtility import get_all_blender_mesh_objects
@@ -121,14 +125,14 @@ class MeshObject(Entity):
         bpy.ops.transform.translate(value=[-bb_center[0], -bb_center[1], -bb_min_z_value])
         bpy.ops.object.mode_set(mode='OBJECT')
         self.deselect()
-        bpy.context.view_layer.update()
 
     def get_bound_box(self, local_coords: bool = False) -> np.ndarray:
         """
         :return: 8x3 array describing the object aligned bounding box coordinates in world coordinates
         """
         if not local_coords:
-            return np.array([self.blender_obj.matrix_world @ Vector(cord) for cord in self.blender_obj.bound_box])
+            local2world = Matrix(self.get_local2world_mat())
+            return np.array([local2world @ Vector(cord) for cord in self.blender_obj.bound_box])
         else:
             return np.array([Vector(cord) for cord in self.blender_obj.bound_box])
 
@@ -218,11 +222,14 @@ class MeshObject(Entity):
         :param temp_dir: The temp dir to use for storing the object files created by v-hacd.
         :param cache_dir: If a directory is given, convex decompositions are stored there named after the meshes hash. If the same mesh is decomposed a second time, the result is loaded from the cache and the actual decomposition is skipped.
         """
+        if platform == "win32":
+            raise Exception("This is currently not supported under Windows")
+
         if temp_dir is None:
             temp_dir = Utility.get_temporary_directory()
 
         # Decompose the object
-        parts = convex_decomposition(self.blender_obj, temp_dir, resolve_path(vhacd_path), cache_dir=resolve_path(cache_dir))
+        parts = convex_decomposition(self, temp_dir, resolve_path(vhacd_path), cache_dir=resolve_path(cache_dir))
         parts = [MeshObject(p) for p in parts]
 
         # Make the convex parts children of this object, enable their rigid body component and hide them
@@ -244,7 +251,7 @@ class MeshObject(Entity):
         :param new_parent: The new parent object.
         """
         self.blender_obj.parent = new_parent.blender_obj
-        self.blender_obj.matrix_parent_inverse = new_parent.blender_obj.matrix_world.inverted()
+        self.blender_obj.matrix_parent_inverse = Matrix(new_parent.get_local2world_mat()).inverted()
 
     def get_parent(self) -> Optional[Entity]:
         """ Returns the parent object.
@@ -606,7 +613,7 @@ def scene_ray_cast(origin: Union[Vector, list, np.ndarray], direction: Union[Vec
             If any object has been hit, the MeshObject otherwise None.
             Some 4x4 matrix.
    """
-    hit, location, normal, index, hit_object, matrix = bpy.context.scene.ray_cast(bpy.context.view_layer.depsgraph,
+    hit, location, normal, index, hit_object, matrix = bpy.context.scene.ray_cast(bpy.context.evaluated_depsgraph_get(),
                                                                                   Vector(origin), Vector(direction),
                                                                                   distance=max_distance)
     if hit_object is not None: