+++ examples/advanced/object_pose_sampling/main.py
@@ -20,7 +20,7 @@ light.set_location([5, -5, 5])
 light.set_energy(1000)
 
 # define the camera intrinsics
-bproc.camera.set_intrinsics_from_blender_params(1, 512, 512, lens_unit="FOV")
+bproc.camera.set_resolution(512, 512)
 
 # read the camera positions file and convert into homogeneous camera-world transformation
 with open(args.camera, "r") as f: